Restructured app so that there was a views folder
