+++ SWRapp/src/components/WeatherAPI/Temp.js
